+++ CHANGELOG.md
@@ -36,6 +36,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - OTU depictions draggable in image matrix
 - Observations depictions in Browse OTU
 - `Ctrl/Alt + V` shortcut for New Collecting event in Comprehensive task [#2248]
+- Zoom button in comprehensive form
 
 ### Changed
 
@@ -62,6 +63,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 - Drag and drop depictions in Image Matrix
 - Georeference error message in comprehensive task [#2222]
 - Number of uses not displayed in Uniquify people task [#2219]
+- SVG Image box in comprehensive [#2262]
 
 [#1934]: https://github.com/SpeciesFileGroup/taxonworks/pull/1934
 [#2169]: https://github.com/SpeciesFileGroup/taxonworks/pull/2169